fix(sonarqube): separate portal lib into src/ and tests/, update sonqarqube config
