+++ portal/common/lib/src/bcs_data_parsing.ts
