+++ portal/common/lib/src/constants.ts
