+++ portal/common/lib/src/custom.d.ts
