+++ portal/common/lib/src/decompress_data.bench.ts
