+++ portal/common/lib/src/decompress_data.ts
