+++ portal/common/lib/src/domain_parsing.ts
