+++ portal/common/lib/src/redirects.ts
