+++ portal/common/lib/src/url_safe_base64.ts
