+++ portal/common/lib/tests/links.test.ts
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { describe, expect, test } from 'vitest';
-import { getObjectIdLink, getBlobIdLink } from './links';
-import { DomainDetails } from './types';
+import { getObjectIdLink, getBlobIdLink } from '../src/links';
+import { DomainDetails } from '../src/types';
 
 const getObjectIdLinkTestCases: [string, DomainDetails | null][] = [
     ["https://example.suiobj.invalid/resource/path", { subdomain: "example", path: "/resource/path" }],
