+++ sonar-project.properties
@@ -6,9 +6,9 @@ sonar.organization=mystenlabs
 sonar.projectName=walrus-sites
 sonar.projectVersion=0.0.1
 
-
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=.
+sonar.sources=portal/common/lib/src,site-builder/src
+sonar.tests=portal/common/tests,site-builder/tests
 
 # Encoding of the source code. Default is default system encoding
 sonar.sourceEncoding=UTF-8
